% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_alp_beta3.R
\name{calc_alp_beta3}
\alias{calc_alp_beta3}
\title{Calculate the omegaVec}
\usage{
calc_alp_beta3(dat, param, omegaVec)
}
\arguments{
\item{dat}{a data matrix that contains all the observations}

\item{param}{a parameter vector that contains alp1, alp2, beta1, beta2 and values of pi}

\item{omegaVec}{the value of the expected....}
}
\value{
a vector that consists the updated parameter values of alp1, alp2, beta1 and beta2
}
\description{
Calculate the omegaVec
}
