% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_EL2.R
\name{calc_EL2}
\alias{calc_EL2}
\title{Calculate the empirical likelihood under SCAR assumption}
\usage{
calc_EL2(dat, param, omegaVec)
}
\arguments{
\item{dat}{a data matrix that contains all the observations}

\item{param}{a parameter vector that contains alp1, alp2, beta1, beta2 and values of pi}

\item{omegaVec}{the vector that determines weighting of the unlabeled positive}
}
\value{
llk the log-likelihood value
}
\description{
Calculate the empirical likelihood under SCAR assumption
}
