% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_omega.R
\name{calc_omega}
\alias{calc_omega}
\title{Calculate the omegaVec}
\usage{
calc_omega(dat, param)
}
\arguments{
\item{dat}{a data matrix that contains all the observations}

\item{param}{a parameter vector that contains alp1, alp2, beta1, beta2 and values of pi}
}
\value{
a vector that consists the values of omega
}
\description{
Calculate the omegaVec
}
