% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.flip_ind}
\alias{.flip_ind}
\title{importFrom("stats", "coef", "filter", "rbinom")}
\usage{
.flip_ind(piVal_true, piVal_hat)
}
\arguments{
\item{piVal_true}{a vector that contains the true pi values}

\item{piVal_hat}{a vector that contains the estimates pi values}
}
\value{
an indicator to show whether or not to flip the pi
}
\description{
importFrom("stats", "coef", "filter", "rbinom")
}
