% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpu3.R
\name{elpu3}
\alias{elpu3}
\title{Calculate the empirical likelihood under the assumption of DET-SAR model}
\usage{
elpu3(dat, param, maxiter = 500, tol = 1e-04, updatePiVal = TRUE, flip = TRUE)
}
\arguments{
\item{dat}{a data matrix that contains all the observations}

\item{param}{a parameter vector that contains alp1, alp2, beta1, beta2 and piVal}

\item{maxiter}{the maximum iteration number}

\item{tol}{the tolerance of the changing in likelihood}

\item{updatePiVal}{whether or not to update piVal}

\item{flip}{whether to flip the pi (proportion) or not}
}
\value{
a list of values: piVal, alp1, alp2, beta1, beta2, llk and iteration number
}
\description{
Calculate the empirical likelihood under the assumption of DET-SAR model
}
