% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dat.R
\name{extract_dat}
\alias{extract_dat}
\title{Extract the data of whether they are labeled or unlabeled}
\usage{
extract_dat(dat, ind_label = 2)
}
\arguments{
\item{dat}{a data matrix that contains all the observations}

\item{ind_label}{an indicator to determine whether to extract the label or unlabeled data}
}
\value{
a matrix with all data that are either labeled or unlabeled
}
\description{
Extract the data of whether they are labeled or unlabeled
}
