% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_pu.R
\name{pred_pu}
\alias{pred_pu}
\title{Calculate the Bayes classifer accurary for the PU data}
\usage{
pred_pu(res_model, dat_new, pu_components)
}
\arguments{
\item{res_model}{the resulting model from the elpu2 or elpu3 functions.}

\item{dat_new}{the testing data set}

\item{pu_components}{this is an indicator about how many components do we assume: either 2 (SCAR) or 3 (SAR)}
}
\value{
The resulting classification rate for the entire data and the unlabelled data only
}
\description{
Calculate the Bayes classifer accurary for the PU data
}
