% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PU_SAR_test.R
\name{PU_SAR_test}
\alias{PU_SAR_test}
\title{Calculate the p-value for the test that $beta_1=0$, which means the the PU data us from the SAR model}
\usage{
PU_SAR_test(dat, param)
}
\arguments{
\item{dat}{a data matrix that contains all the observations}

\item{param}{a parameter vector that contains alp1, alp2, beta1, beta2 and piVal}
}
\value{
a list of values: piVal, alp1, alp2, beta1, beta2, llk and iteration number
importFrom("stats", "pchisq")
}
\description{
Calculate the p-value for the test that $beta_1=0$, which means the the PU data us from the SAR model
}
