% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUfit.R
\name{PUfit}
\alias{PUfit}
\title{PU fitting with four models (SETMnull;SETMfull;DETMnull;DETMfull)}
\usage{
PUfit(posdata, negdata, pival, alp1, beta1, alp2, beta2)
}
\arguments{
\item{posdata}{n by p feature matrix for the source (positive) data in PU data.}

\item{negdata}{m by p feature matrix for the target (unlabled) data in PU data.}

\item{pival}{initial value of the mixture proportion of positive data in the target dataset, is set for the EM algorithm.}

\item{alp1}{initial value of alpha_1 in EM algorithm}

\item{beta1}{A vector of initial value of beta_1 in EM algorithm}

\item{alp2}{initial value of alpha_2 in EM algorithm}

\item{beta2}{A vector of initial value of beta_2 in EM algorithm}
}
\value{
A list including four models with their mixture proportion estimation, parameter estimation, iterations, likelihood and penalized likelihood.
}
\description{
PU fitting with four models (SETMnull;SETMfull;DETMnull;DETMfull)
}
