% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SETMfull.R
\name{SETMfull}
\alias{SETMfull}
\title{SETM fitting with full parameters}
\usage{
SETMfull(posdata, negdata, pival, alp, beta, maxite = 1000, eps = 1e-04)
}
\arguments{
\item{posdata}{n by p feature matrix for the source (positive) data in PU data.}

\item{negdata}{m by p feature matrix for the target (unlabled) data in PU data.}

\item{pival}{initial value of the mixture proportion of positive data in the target dataset, is set for the EM algorithm.}

\item{alp}{initial value of alpha of SETM in EM algorithm}

\item{beta}{A vector of initial value of beta of SETM in EM algorithm}

\item{maxite}{maximum iteration number of EM algorithm}

\item{eps}{tolerance value about stopping the algorithm when the increment in the log-EL after an iteration is no greater than, e.g. 1e-4.}
}
\value{
A list including mixture proportion estimation, parameter estimation, iterations, likelihood and penalized likelihood in SETM.
}
\description{
SETM fitting with full parameters
}
