% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STEMnull.R
\name{SETMnull}
\alias{SETMnull}
\title{SETM fitting with parameters given mixture proportion of positive data in the target dataset}
\usage{
SETMnull(posdata, negdata, pival, alp, beta, maxite = 1000, eps = 1e-04)
}
\arguments{
\item{posdata}{n by p feature matrix for the source (positive) data in PU data.}

\item{negdata}{m by p feature matrix for the target (unlabled) data in PU data.}

\item{pival}{fixed mixture proportion of positive data in the target dataset, is set for the EM algorithm.}

\item{alp}{initial value of alpha of SETM in EM algorithm}

\item{beta}{A vector of initial value of beta of SETM in EM algorithm}

\item{maxite}{maximum iteration number of EM algorithm}

\item{eps}{tolerance value about stopping the algorithm when the increment in the log-EL after an iteration is no greater than, e.g. 1e-4.}
}
\value{
A list including fixed mixture proportion estimation, parameter estimation, iterations, log-likelihood and penalized log-likelihood in SETM.
}
\description{
SETM fitting with parameters given mixture proportion of positive data in the target dataset
}
